requirements.txt
flask==3.0.0
groq==0.11.0
python-dotenv==1.0.0
moviepy==1.0.3
httpx==0.27.0
instaloader==4.10.3
yt-dlp==2024.12.23
beautifulsoup4==4.12.3
requests==2.31.0
flask-socketio==5.3.6
python-engineio==4.9.0
python-socketio==5.11.1
brotli==1.1.0
